+++ lock-impl/src/main/java/com/palantir/lock/impl/TooManyRequestsException.java
@@ -15,7 +15,7 @@
  */
 package com.palantir.lock.impl;
 
-public class TooManyRequestsException extends RuntimeException{
+public class TooManyRequestsException extends RuntimeException {
     public TooManyRequestsException(String message) {
         super(message);
     }